separated server and app, setup for writing tests
